+++ stingraykit/io/BufferedPipe.h
@@ -28,8 +28,8 @@ namespace stingray
 	public:
 		explicit BufferedPipe(const IPipePtr& pipe, size_t bufferSize = DefaultBufferSize);
 
-		virtual size_t Read(ByteData data, const ICancellationToken& token);
-		virtual size_t Write(ConstByteData data, const ICancellationToken& token)
+		virtual u64 Read(ByteData data, const ICancellationToken& token);
+		virtual u64 Write(ConstByteData data, const ICancellationToken& token)
 		{ return _pipe->Write(data, token); }
 	};
